+++ api/client/src/main/java/org/projectnessie/client/builder/BaseAssignReferenceBuilder.java
@@ -13,23 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.projectnessie.client.http.v2api;
+package org.projectnessie.client.builder;
 
-import org.projectnessie.client.api.AssignBranchBuilder;
-import org.projectnessie.client.http.HttpClient;
-import org.projectnessie.model.Branch;
-import org.projectnessie.model.Reference.ReferenceType;
+import org.projectnessie.model.Reference;
 
-final class HttpAssignBranch extends BaseHttpAssignReference<Branch, AssignBranchBuilder>
-    implements AssignBranchBuilder {
+public abstract class BaseAssignReferenceBuilder<R> extends BaseChangeReferenceBuilder<R> {
 
-  HttpAssignBranch(HttpClient client) {
-    super(client);
-    refType(ReferenceType.BRANCH);
-  }
+  protected Reference assignTo;
 
-  @Override
-  public AssignBranchBuilder branchName(String branchName) {
-    return refName(branchName);
+  @SuppressWarnings("unchecked")
+  public R assignTo(Reference assignTo) {
+    this.assignTo = assignTo;
+    return (R) this;
   }
 }